refactor: separate custom marshalling logic from model
